requirements.txt
Flask==0.12.3
keras==2.2.4
tensorflow==1.13.1
pillow==5.4.1
numpy==1.16.2
opencv-python==4.0.0.21